remove buttons from the center, removed unncessary ListemItemButtons on the sidebar,moved functionality from those bottons to the timer's play button
