--- src/__tests__/App.test.tsx
@@ -1,9 +1,0 @@
-import '@testing-library/jest-dom';
-import { render } from '@testing-library/react';
-import App from '../renderer/App';
-
-describe('App', () => {
-  it('should render', () => {
-    expect(render(<App />)).toBeTruthy();
-  });
-});
