+++ src/renderer/components/timer/TimerDisplay.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useTimer } from '../../contexts/TimerContext';
 
 import PlayCircleFilledWhiteIcon from '@mui/icons-material/PlayCircleFilledWhite';
@@ -10,20 +10,33 @@ import PlatformToggle from '../platform-toggle/PlatformToggle';
 
 export default function TimerDisplay() {
   const { elapsedTime, formatTime, isRunning, setIsRunning } = useTimer();
-
+  const [processRunning, setProcessRunning] = useState<boolean>(false);
   return (
     <div className="timer-container">
       <div className="timer-content">
         {isRunning ? (
           <StopCircleIcon
-            sx={{ width: '35px', height: '35px' }}
+            sx={{ width: '35px', height: '35px', cursor: 'pointer' }}
             className="timer-icon-stop"
           />
         ) : (
-          <PlayCircleFilledWhiteIcon
-            sx={{ width: '35px', height: '35px' }}
-            className="timer-icon-play"
-          />
+          <div
+            onClick={() => {
+              if (processRunning) {
+                const userConfirmed = window.confirm(
+                  'The process is already running. Do you want to start it again?',
+                );
+                if (!userConfirmed) return;
+              }
+              window.electron.ipcRenderer.sendMessage('run-python-script', []);
+              setProcessRunning(true);
+            }}
+          >
+            <PlayCircleFilledWhiteIcon
+              sx={{ width: '35px', height: '35px', cursor: 'pointer' }}
+              className="timer-icon-play"
+            />
+          </div>
         )}
 
         <div className="timer-text">{formatTime(elapsedTime)}</div>
